setup X-Digest middleware

+++ internal/models/wallet.go
@@ -9,7 +9,7 @@ type Wallet struct {
 	ID        uuid.UUID `json:"id" gorm:"primaryKey;type:uuid;default:gen_random_uuid()"`
 	UserID    uuid.UUID `json:"user_id"`
 	Type      string    `json:"type" gorm:"default:non-identified"`
-	Balance   float64   `json:"balance"`
+	Balance   uint64    `json:"balance"`
 	CreatedAt time.Time `json:"created_at"`
 	UpdatedAt time.Time `json:"updated_at"`
